client-api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
scapy = "^2.5.0"
flask-cors = "^4.0.1"
grpcio = "^1.63.0"
grpcio-tools = "^1.63.0"
pypi = "^2.1"
requests = "^2.31.0"

[tool.poetry.scripts]
start = "api.__init__:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
